pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "nltk==3.9.1",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "sentence-transformers>=5.1.2",
]
